week_15/day_5/CryptoCurrenciesProject-copie/requirements.txt
alembic==1.8.0
Babel==2.10.3
Flask==2.1.2
Flask_Bcrypt==1.0.1
Flask_Login==0.6.1
Flask_Mail==0.9.1
Flask_Migrate==3.1.0
Flask_SQLAlchemy==2.5.1
Flask_WTF==1.0.1
itsdangerous==2.1.2
PyJWT==2.4.0
python-dotenv==0.20.0
python_dateutil==2.8.2
requests==2.27.1
schedule==1.1.0
SQLAlchemy==1.4.37
WTForms==3.0.1
email-validator==1.2.1
psycopg2==2.9.3
gunicorn